Add xrays
