+++ frontend/src/metabase/home/landing/components/LandingApp/LandingApp.jsx
@@ -1,27 +1,39 @@
 import React from "react";
 import PropTypes from "prop-types";
+import GreetingSection from "../GreetingSection";
+import OurDataSection from "../OurDataSection";
+import XraySection from "../XraySection";
 import { LandingRoot } from "./LandingApp.styled";
 
 const propTypes = {
-  GreetingSection: PropTypes.elementType,
-  XraySection: PropTypes.elementType.isRequired,
-  OurDataSection: PropTypes.elementType.isRequired,
+  greeting: PropTypes.string,
+  databases: PropTypes.array,
+  databaseCandidates: PropTypes.array,
+  isAdmin: PropTypes.bool,
   showXrays: PropTypes.bool,
   showOurData: PropTypes.bool,
 };
 
 const LandingApp = ({
-  GreetingSection,
-  XraySection,
-  OurDataSection,
+  greeting,
+  databases,
+  databaseCandidates,
+  isAdmin,
   showXrays,
   showOurData,
 }) => {
   return (
     <LandingRoot>
-      <GreetingSection />
-      {showXrays && <XraySection />}
-      {showOurData && <OurDataSection />}
+      <GreetingSection greeting={greeting} />
+      {showXrays && (
+        <XraySection
+          databaseCandidates={databaseCandidates}
+          isAdmin={isAdmin}
+        />
+      )}
+      {showOurData && (
+        <OurDataSection databases={databases} isAdmin={isAdmin} />
+      )}
     </LandingRoot>
   );
 };
