--- frontend/src/metabase/home/landing/containers/GreetingSection/GreetingSection.jsx
@@ -1,9 +1,0 @@
-import { connect } from "react-redux";
-import GreetingSection from "../../components/GreetingSection";
-import { getGreeting } from "../../selectors";
-
-const mapStateToProps = state => ({
-  greeting: getGreeting(state),
-});
-
-export default connect(mapStateToProps)(GreetingSection);
