--- frontend/src/metabase/home/landing/containers/GreetingSection/index.js
@@ -1,1 +1,0 @@
-export { default } from "./GreetingSection";
