--- frontend/src/metabase/home/landing/containers/OurDataSection/index.js
@@ -1,1 +1,0 @@
-export { default } from "./OurDataSection";
