--- frontend/src/metabase/home/landing/containers/XraySection/index.js
@@ -1,1 +1,0 @@
-export { default } from "./XraySection";
